[WIP] Begin aggressive integration with HTMLBars
This commit starts to integrate with the more aggressive HTMLBars
changes. In particular, HTMLBars now takes more control over the
rendering process, and delegates to Ember as it proceeds.

Ember is responsible for managing the template scope, while HTMLBars is
responsible for walking a tree of render nodes and updating them. One
nice effect of the HTMLBars concept of “scope” is evident in the
bind-self hook, which is now able to encapsulate the concept of Ember
“keywords” in the same local variable scope in HTMLBars that is used to
look up block arguments.

You can also take a look at the if_unless helper file, which can now
be written as a simple imperative function that is completely isolated
from concerns about the view hierarchy. Instead of creating a new view,
Ember can now make use of the render nodes and scope that are provided
by HTMLBars.

This is a fairly early attempt at integration, mostly focusing on the
tests in ember-htmlbars, but so far so good.